Пятое дз выполнено. Отправка на проверку
